SHMRK-3544 - Add Paging Functionality to Messages Threads List

+++ CarePay/CarePayPatient/src/main/java/com/carecloud/carepay/patient/messages/MessageNavigationCallback.java
@@ -14,7 +14,7 @@ public interface MessageNavigationCallback {
 
     void displayThreadMessages(Messages.Reply thread);
 
-    void getMessageThreads(int page, int size);
+    void getMessageThreads(long page, long size);
 
     void getThreadMessages(Messages.Reply thread);
